refclock_gpsd: stop some unsigned/signed shuffling.
Fixes a warning on 32-bit RasPi.

+++ ntpd/refclock_gpsd.c
@@ -2050,7 +2050,7 @@ convert_ascii_time(
 	char           *ep;
 	struct tm       gd;
 	struct timespec ts;
-	uint32_t        dw;
+	long   dw;
 
 	/* Use 'strptime' to take the brunt of the work, then parse
 	 * the fractional part manually, starting with a digit weight of
